% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.starter-package.R
\docType{package}
\name{bioinfo.starter-package}
\alias{bioinfo.starter}
\alias{bioinfo.starter-package}
\title{bioinfo.starter: A template for bioinformatic and ecological analysis of metabarcoding}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/adrientaudiere/bioinfo.starter/issues}
}

}
\author{
\strong{Maintainer}: Adrien Taudière \email{adrien.taudiere@zaclys.net} (\href{https://orcid.org/0000-0003-1088-1182}{ORCID}) [copyright holder]

}
\keyword{internal}
